+++ src/main/java/woowacourse/shoppingcart/domain/Password.java
@@ -1,7 +1,5 @@
 package woowacourse.shoppingcart.domain;
 
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.Objects;
 import java.util.regex.Pattern;
 import woowacourse.exception.LoginException;
@@ -9,8 +7,9 @@ import woowacourse.exception.dto.ErrorResponse;
 
 public class Password {
 
-    private String password;
     private static final Pattern PATTERN = Pattern.compile("^.*(?=^.{8,12}$)(?=.*\\d)(?=.*[a-zA-Z])(?=.*[!@#$%^&+=]).*$");
+
+    private final String password;
 
     private Password(String password) {
         this.password = password;
@@ -18,7 +17,7 @@ public class Password {
 
     public static Password from(String password) {
         validatePassword(password);
-        String encryptedPassword = encrypt(password);
+        String encryptedPassword = PasswordEncryptUtil.encrypt(password);
         return new Password(encryptedPassword);
     }
 
@@ -26,28 +25,6 @@ public class Password {
         if (!PATTERN.matcher(password).matches()) {
             throw new LoginException("비밀번호 규약이 맞지 않습니다", ErrorResponse.INVALID_PASSWORD);
         }
-    }
-
-    private static String encrypt(String password) {
-        try {
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            md.update(password.getBytes());
-            return bytesToHex(md.digest());
-        } catch (NoSuchAlgorithmException e) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private static String bytesToHex(byte[] bytes) {
-        StringBuilder builder = new StringBuilder();
-        for (byte b : bytes) {
-            builder.append(String.format("%02x", b));
-        }
-        return builder.toString();
-    }
-
-    public static boolean isSameEncryptedPassword(String encryptedPassword, String password) {
-        return encryptedPassword.equals(encrypt(password));
     }
 
     public String getPassword() {
